fix(zeb-295): engine-orchestration completeness for ballot-secret tally
Four findings from PR #155 reviewers (CodeAnt critical, CodeRabbit major,
Cursor #1, Qodo #5) — all four converge on engine-side completeness gaps
that left se-mode polls unable to finalize even with all the apply-time
math from C1+C2 landed. Without this batch, a se-mode poll could pass
every verify rule, accept every ballot, and still stall in Ratification
permanently because no node ever published kd=ts and no non-committee
node ever finalized kd=rs.

Finding 1 (CodeAnt critical — incomplete implementation).
`maybe_emit_tally_share` was gated correctly but had no emission body —
just a `tracing::debug!` placeholder with "deferred to follow-up". Now
fully implemented: snapshot poll-state under lock, resolve the local
FROST `KeyPackage` from `DfrostLogRegistry`, derive `x_i` via
`signing_share_as_scalar`, compute `d_i + DLEQ proof` for each of the
`n + C(n,2)` homomorphic aggregates, and publish a signed
`TallySharePayload` envelope. `n` derivation matches the C2 canonical
ratification-ordering computation. Bails silently on every missing
prerequisite (no committee, no local key, aggregate decode failure)
so future applies can retry as prerequisites arrive.

Finding 2 (Cursor #1 + Qodo #5 + CodeRabbit — oracle never retried).
`DfrostLogCommitteeOracle` is installed exactly once at PollCreate
apply time. If no DKG has completed yet at that moment, the poll
permanently binds to `NullCommitteeOracle` — every kd=rb NIZK verify
silently fails (no joint Y available), every kd=ts T1/T2 verify fails
(no Y_i). Implementing Option A (hard gate at PollCreate IPC, matches
`voting_cast_ratification_ballot`'s existing fail-fast pattern):
`voting_create_tier3_proposal` now refuses `privacy_mode="se"` when
`DfrostLogRegistry::get(community).latest_committee_epoch()` returns
`None`, with the error message "se-mode polls require an active
community committee; complete DKG ceremony first".

Finding 3 (CodeRabbit major — inbound kd=ts doesn't trigger kd=rs).
`maybe_emit_tier3_result_secret` (renamed `try_finalize_secret_tally`)
was only called from `maybe_trigger_engine_auto_orchestration`, which
itself only fires on the OUTBOUND `publish_event` path. A node that
is NOT in the committee never publishes kd=ts itself, so its outbound
cascade never runs — it would observe ≥t kd=ts from peers, hold all
the inputs `recover_secret_tally` needs, and never finalize its own
log. Extract the recover+publish body into `try_finalize_secret_tally`
and invoke it from BOTH paths: the existing outbound cascade AND a
new gate at the end of `process_inbound_dispatch` that fires when an
inbound Tier::Sortition event is kd=ts or kd=cl. Lagrange invariance
+ canonical share-subset selection guarantees bit-identical kd=rs
envelopes across replicas, so the apply-time LWW gate cleanly resolves
the inevitable race.

Finding 4 (CodeAnt critical — epoch u32 vs u64 mismatch).
`TallySharePayload.committee_epoch` was `u32` but
`Tier3PollMeta.community_epoch` and `committee_state.current_epoch`
are both `u64`. Cross-module type-contract drift that would silently
truncate once CHURP rotations exceeded `u32::MAX`. Worse, even before
that horizon the CBOR encoding length differs between u32 and u64, so
a future widening would break compat with already-shipped envelopes —
fixing this now, before Phase 6 ships, is the only race-free option.
Touched points: `TallySharePayload`, `SecretTallyState::tally_shares`
key, `CommitteeOracle` trait sigs, `DfrostLogCommitteeOracle::epoch`,
`DfrostLogEngine::{latest_committee_epoch, committee_snapshot_at_epoch}`,
`voting-tier3-tally-share-applied` Tauri event payload + TS type, all
test helper signatures + mock oracle impls. Wire-format fixtures will
regenerate on the next REGENERATE_VOTING_TIER3_SECRET_FIXTURES run.

Tests.
New `community_voting_tier3_secret_kd_ts_emission_integration.rs`
drives a real 3-party 2-of-3 FROST DKG, wires the full DfrostLogEngine
+ VotingLogEngine pair, seeds a se-mode poll past kd=cl with a
production `DfrostLogCommitteeOracle` installed, and asserts:
  - exactly one kd=ts envelope is broadcast on `publisher_tx`
  - the envelope's `committee_epoch + entries.len()` match the
    expected canonical ratification ordering
  - the emitted envelope round-trips through `apply_event` (T1/T2
    verify accepts it) — soundness check that emission is bit-exact
    with the apply path's expectations
  - re-invoking the hook does NOT publish a second kd=ts
    (already-emitted gate)
  - a non-committee node (signing-owner not in `verifying_shares`)
    silently bails — no kd=ts emission

A `#[cfg(any(test, feature = "test-fixtures"))]` test seam exposes
`maybe_emit_tally_share` so the test can drive the hook in isolation
without scaffolding a full chain of kd=ss → kd=dc → kd=rb → kd=cl
signed events.

+++ src-tauri/src/community_voting_core.rs
@@ -319,8 +319,19 @@ pub struct TallySharePayload {
     #[serde(rename = "pi")]
     pub poll_id: PollId,
     /// CHURP rotation generation. Shares from different epochs cannot be mixed.
+    ///
+    /// `u64` to match `DfrostLogEngine::current_epoch` (the DKG-side source of
+    /// truth) and `Tier3PollMeta::community_epoch`. The earlier `u32` shape
+    /// was a cross-module type-contract drift (CodeAnt PR #155 critical):
+    /// if the CHURP rotation counter ever exceeds `u32::MAX` (4 billion
+    /// rotations) the tally-share envelope's epoch silently truncates and
+    /// no longer matches the dfrost log's epoch, so the apply path drops
+    /// every kd=ts. Even below that horizon, the encoded-as-u32 wire byte
+    /// length differs from the encoded-as-u64 length, so a future widening
+    /// would break compatibility with already-shipped envelopes — fixing
+    /// this NOW, before Phase 6 ships, is the only race-free option.
     #[serde(rename = "ce")]
-    pub committee_epoch: u32,
+    pub committee_epoch: u64,
     /// `n + C(n,2)` entries: candidate score-sum entries first, then indicator-sum
     /// entries in unordered-pair lexicographic order. Vec (not fixed array) because
     /// `n` is per-poll and only known at apply time.
